Fixed screen saver handling
- Coniguration of screen saver is now part of interface IScreenControl, configuration of screen saver uses that interface
- Improved screen saver configuration texts

+++ MediaPortal/Source/UI/SkinEngine/Settings/ScreenSaverSettings.cs
@@ -32,6 +32,6 @@ namespace MediaPortal.UI.SkinEngine.Settings
     public bool ScreenSaverEnabled { get; set; }
 
     [Setting(SettingScope.User, 5)]
-    public int ScreenSaverTimeoutMin { get; set; }
+    public double ScreenSaverTimeoutMin { get; set; }
   }
 }
